Apply static analysis levels 4 and 5

+++ app/Models/Leaderboard.php
@@ -37,7 +37,7 @@ class Leaderboard extends Model
   */
   public function getRatingAttribute($value)
   {
-    return number_format($value, 2);
+    return number_format((float) $value, 2);
   }
 
   /**
@@ -48,7 +48,7 @@ class Leaderboard extends Model
   */
   public function getWinRateAttribute($value)
   {
-    return number_format($value, 2);
+    return number_format((float) $value, 2);
   }
 
   /**
@@ -59,6 +59,6 @@ class Leaderboard extends Model
  */
   public function getMmrAttribute($value)
   {
-      return round(1800 + 40 * (int)$value);
+      return round(1800 + 40 * (float) $value);
   }
 }
